Move tests for old API into their own file.
