+++ test/integration/bills_test.rb
@@ -1,32 +1,6 @@
 require 'sunlight/congress'
 require 'webmock/minitest'
 
-# This stuff is behavior from initial development. It can be removed with a
-# major version bump.
-class TestIntegrationCongressBillOldAPI < MiniTest::Unit::TestCase
-  def setup
-    Sunlight::Congress.api_key = "thisismykey"
-  end
-
-  def test_bills_by_title
-    stub_request(:get, "http://congress.api.sunlightfoundation.com/bills/search?apikey=thisismykey&query=Health").
-      to_return(:status => 200, :body => '{"results":[{"title":"Awesome Health Care Bill"}]}', :headers => {})
-
-    title = "Health"
-    bills = Sunlight::Congress::Bill.by_title(title)
-
-    assert_equal "Awesome Health Care Bill", bills.first.title
-  end
-
-  def test_bills_by_id
-    stub_request(:get, "http://congress.api.sunlightfoundation.com/bills?bill_id=s668-113&apikey=thisismykey").
-          to_return(:status => 200, :body => '{"results":[{"bill_id":"s668-113"}]}')
-    bill = Sunlight::Congress::Bill.by_bill_id("s668-113")
-
-    assert_equal "s668-113", bill.bill_id
-  end
-end
-
 class TestIntegrationCongressBill < MiniTest::Unit::TestCase
   def setup
     @api = Sunlight::Congress.new("thisismykey")
